reverse image search

+++ plugins/upscale.py
@@ -1,5 +1,3 @@
-# Copyright 2023 Qewertyy, MIT License
-
 from pyrogram import Client, filters, types as t
 from Utils import getFile, UpscaleImages
 import os
@@ -7,8 +5,10 @@ import os
 @Client.on_message(filters.command(["upscale"]))
 async def upscaleImages(_, message):
     file = await getFile(message)
+    if file == 1:
+       return await message.reply_text("File size is large")
     if file is None:
-        return await message.reply_text("Replay to an image?")
+       return await message.reply_text("Replay to an image?")
     msg = await message.reply("wait a min...")
     imageBytes = open(file,"rb").read()
     os.remove(file)
